:recycle: 使用wire重构代码

+++ src/entity/user.go
@@ -1,7 +1,7 @@
 package entity
 
 type User struct {
-	Uid      int    `json:"uid"`
+	Uid      int64  `json:"uid"`
 	Nickname string `json:"nickname"`
 	Password string `json:"password"`
 }
